Rewrite the implementation completely
Instead of using the filespec in the autocmds, we define a single buffer-scoped autocmd, and keep the filespecs in the b:DuplicateWrite variable. This makes it easier to avoid adding the same cascaded write target twice. The duplication is automatically cleared on :bdelete, so we don't need the configurable behavior any more. It also means the list functions become more involved, but this also allows us to improve on the output format.