BUG: A netrw target (e.g. scp://hostname/path) causes the autocmds to get lost, because it temporarily :setlocal nobuflisted, and that triggers the BufDelete event
Wrap its action in g:DuplicateWrite_Working check to prevent that.
Invoke (now public) DuplicateWrite#EnsureAutocmd() after the write to redefine the autocmds if necessary, and warn about this (in verbose mode).
FIX: Cleanup of autocmds may not apply after :bdelete, as the target buffer may not be the current one. Encode the buffer number in the action.